Structural size grinding.
Down to 713 cells by
- Allowing more flop enable signals to be inferred
- Removing some unnecessary mux cases
- Random magic

The implementation is down to 95 lines of Haskell. I haven't gone crazy
on the line golfing, it's fairly idiomatic. Compared to the original J1
Verilog sources, it's missing the inferred RAMs, but adds in a bunch of
language pragmas and imports, so I feel like this is definitely a
compactness win for Clash.