Remove load bit, use J1-style loads.
When the load bit was set, the Tmux field was effectively ignored. Or
more to the point, it was processed in the first cycle, and then
immediately overwritten. This seemed dumb.

Loads are now triggered when Tmux is 12 (the J1's [T] operation). Given
the four-bit Tmux field, this comparator costs a single LUT4.

Bit 4 of the ALU instruction is now available again.

I believe that, at this point, CFM is back to being J1-compatible in
everything except timing.