Make demo fully synthesizable.
We can now get a timing report. The current design is just north of
50MHz on the HX8k device on the icoboard. The critical path,
surprisingly, is sign extension of the magnitude comparator to all bits
of T.

The constraints are for generating a (boring) pattern on the LEDs on the
icoboard. However, we don't yet meet timing for the icoboard's fast
oscillator, and I haven't instantiated a PLL.