Fix subtle bug in RAM stack implementation.
The RAM stacks were implemented as read-before-write, meaning a value
pushed in cycle N was not available in cycle N+1. Remarkably, my UART
code just happened to work despite this -- but optimizations caused it
to break, leading me to this bug.

The RAMs are now filled with exploding error values for simulation, to
catch this sort of mistake in the future.