Basic interrupt support.
I've done something odd here. The CPU is still ignorant of interrupts;
the only CPU change was exposing the Fetch signal in a previous commit.
The interrupt controller operates by interposing between the CPU and
RAM, and altering fetched instructions.

This works because a single-word call instruction can reach the low
portion of RAM, where vectors live, and we don't need to bank any
registers on ISR entry.

Separately, the interrupt controller exposes a write-sensitive trigger
register for re-enabling interrupts. It disregards the actual value
written. This enables a subtle fused instruction sequence to write an
arbitrary value to the trigger, drop the trigger address, and return
from the ISR, in a single cycle -- without the need for a RETI
instruction.