GPIO: describe both ports as Moore machines.
This reduces a bunch of functor noise, which actually caused me to
notice and fix a bug in the input port. (It was only asserting irq for a
single cycle. This happens to be enough, but it's wrong.)

Moore machines are a nice description of I/O devices on this core,
because a Moore machine can't have a combinational path from the bus
request to the response. (Mealy machines can.)