icoboard: boot from SPI flash.
This change reduces the BRAM on the Icoboard to a 256-word boot ROM
(plus the BRAMs used for stack and graphics). The ROM contains a small
program that loads memory from SPI flash, starting 256kiB in (and thus
past the FPGA bitstream). It then switches itself out, so the system
runs entirely out of external SRAM.

The reduction in BRAM addressing circuitry has significantly improved
timing to around 47MHz average.