Significantly improve bootstrap speed.
This is well over 100x faster in practice and makes bootstrapping a
non-event. Rather than pretending the system is remote and communicating
with a debug stub through H2T, this directly mucks about in the machine
state.

This reduces the number of cycles emulated by 112x, but I think the
approach I'm using to drive the emulator is also significantly faster
(not using Chan).