RTL/IOBus: hit mealyp/moorep with INLINE hammer.
I'm annoyed that I had to do this explicitly.

In Clash's lowering of these combinators, the `def` value wound up
appearing as an input at the Verilog level, generated at the
instantiation site (and combined with the 0 default for the response
mux). This meant the reset value for the registers generated by the
Moore machine depended on an input. An input that, in practice, was
constant, but an input nonetheless.

This got a warning from Yosys and produced results that appear to be
flaky, albeit for reasons I don't understand.