Reduce speed to 36MHz.
Something in the circuit is marginal. At 40MHz, small changes in the RTL
will cause the system to start failing. Not *all* of the system -- it's
almost always around the small memories implemented using flops with
asynchronous read, such as the video palette and timing registers.

Timing sez 46-50MHz depending on PnR results.

Since the bulk of the system is expressed as Mealy/Moore machines at the
Clash level, I'm not sure what's causing this discrepancy. Fortunately,
there's an 800x600 standard mode at 36MHz.