Results of first self-hosting sprint.
bf/meta.fs is a block-formatted file (as accepted by blocktool)
containing a rough port of parts of the Haskell bootstrapper. At this
point it can compile the beginning of the bsforth sources. To operate,
load into bsforth and run

    begin-bootstrap
    n m <thru>
    end-bootstrap

where n m is a range of blocks you wish to use to bootstrap.